Initialize persistor on the StoreAccessor class
Inject store from the app file after initialization
