+++ __tests__/App.spec.js
@@ -12,9 +12,6 @@ import { UnconnectedApp } from '../App';
 
 jest.mock('../app/Entry', () => ({ Entry: 'Entry' }));
 
-// We're using the unconnected App, so omit creating a store.
-jest.mock('../app/store', () => ({ createPersistedStore: () => ({}) }));
-
 it('renders correctly', () => {
   const { asJSON } = render(<UnconnectedApp />);
