+++ app/bt/index.ts
@@ -21,7 +21,6 @@ const btStrategy: TracingStrategy = {
   affectedUserFlow: AffectedUserFlow,
   assets: btAssets,
   useCopy: useBTCopyContent,
-  extraMiddleware: [],
 };
 
 export { BTNativeModule };
