+++ app/factories/tracingStrategy.tsx
@@ -24,7 +24,6 @@ export default Factory.define<TracingStrategy>(() => ({
   affectedUserFlow: AffectedUserFlow,
   assets: testStrategyAssets,
   useCopy: () => testStrategyCopy,
-  extraMiddleware: [],
 }));
 
 const PermissionsProvider = ({
