+++ app/tracingStrategy.ts
@@ -1,5 +1,4 @@
 import { ImageSourcePropType } from 'react-native';
-import { Dispatch, Middleware } from 'redux';
 
 import { TFunction } from 'i18next';
 
@@ -13,7 +12,6 @@ export interface TracingStrategy {
   affectedUserFlow: () => JSX.Element;
   assets: StrategyAssets;
   useCopy: StrategyCopyContentHook;
-  extraMiddleware: Middleware<Dispatch>[];
 }
 
 export interface StrategyAssets {
